requirements.txt
Flask~=3.1.0
Werkzeug~=3.1.3
tensorflow~=2.15.0
tensorflow-hub
Pillow~=10.4.0
numpy~=1.26.4
requests~=2.32.3
watchdog~=4.0.2
python-dotenv~=1.0.1